Project 5b: Using an EOF Controlled While Loop
This project will use an EOF controlled while loop to access
data from a file. The program will calculate the weekly gross pay based
on hours and rate for hourly employees, those that have an 'h' in the
classification field or on the value in the salary field for those with
a
classification of 's'. Hourly employees with more than 40 hours worked
in a
week will be paid at time and 1/2 for all hours over 40 worked in the
week.